rag-basic/requirements.txt
Flask
transformers
faiss-cpu
torch
datasets
sentence-transformers